api_tp_2/requirement.txt
Django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
psycopg2-binary
pandas
pyarrow